requirements.txt
alembic==1.13.1
cx_Oracle==8.3.0
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
typing_extensions==4.12.1
